Convert Settings index file to ts
